refactor: make custom hook useCurrentPath
